refactor(core): extract maintenance/low-signal.ts — low-signal memory deactivation
